feat(frontend): airdrop completed messages
