+++ .env.example
@@ -16,5 +16,6 @@ VITE_OISY_URL=https://oisy.com
 VITE_OISY_ICON=https://oisy.com/favicons/icon-512x512.png
 VITE_OISY_REPO_URL=https://github.com/dfinity/oisy-wallet
 VITE_AIRDROP=false
+VITE_AIRDROP_COMPLETED=false
 VITE_ERC20_ICP_SYMBOL=ckICP
 VITE_ERC20_ICP_REPO_URL=https://github.com/dfinity/erc20-icp
